requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
dataclasses==0.6
filelock==3.0.12
Flask==1.1.2
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
-e git+https://github.com/OpenNMT/OpenNMT-py.git@1b1e6c1ce5b29d09e87881797dce786148f51090#egg=OpenNMT_py
packaging==20.4
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyonmttok==1.22.2
pyparsing==2.4.7
pyre-check==0.9.6
pyre-extensions==0.0.27
pywatchman==1.4.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.91
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.0
torchtext==0.4.0
torchvision==0.8.1
tqdm==4.52.0
transformers==4.2.2
typing-extensions==3.7.4.3
urllib3==1.26.2
waitress==1.4.4
Werkzeug==1.0.1
plotly
kaleido

